# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SLM-CV OpenAPI Documentation",
        "description": "This is the OpenAPI specification for the Controlled Vocabulary.",
        "contact": {
            "name": "SLM-CV Support",
            "url": "http://localhost/api/v1",
            "email": "support@igs.org"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/license/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "SLM-CV API Server"
        }
    ],
    "paths": {
        "/terms": {
            "get": {
                "tags": [
                    "terms"
                ],
                "summary": "List terms",
                "description": "Display a listing of the resource.",
                "operationId": "indexTerms",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/search"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/vnd.ld+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Term"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Term"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/terms/{uuid}": {
            "get": {
                "tags": [
                    "terms"
                ],
                "summary": "Show term",
                "description": "Display the specified resource.",
                "operationId": "showTerm",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "uuid",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/vnd.ld+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Term"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Term"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/vocabularies/{uuid}/terms": {
            "get": {
                "tags": [
                    "terms",
                    "vocabularies"
                ],
                "summary": "List terms by vocabulary",
                "description": "Display a listing of the resource.",
                "operationId": "indexTermVocabularies",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "uuid",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/search"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/vnd.ld+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Term"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Term"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/vocabularies": {
            "get": {
                "tags": [
                    "vocabularies"
                ],
                "summary": "List vocabularies",
                "description": "Display a listing of the resource.",
                "operationId": "indexVocabularies",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/search"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/vnd.ld+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vocabulary"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Vocabulary"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/vocabularies/{uuid}": {
            "get": {
                "tags": [
                    "vocabularies"
                ],
                "summary": "Show vocabulary",
                "description": "Display the specified resource.",
                "operationId": "showVocabulary",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "uuid",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/vnd.ld+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vocabulary"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Vocabulary"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Term": {
                "description": "Term schema",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "title": "Term name",
                        "description": "Name of the term",
                        "type": "string",
                        "format": "string",
                        "nullable": false
                    },
                    "definition": {
                        "title": "Term definition",
                        "description": "Definition of the term",
                        "type": "string",
                        "format": "string",
                        "nullable": true
                    },
                    "createdAt": {
                        "title": "Created at",
                        "description": "Timestamp of creation",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-08-23 14:39:37",
                        "nullable": true
                    },
                    "updatedAt": {
                        "title": "Updated at",
                        "description": "Timestamp of last update",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-08-23 14:39:37",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Vocabulary": {
                "description": "Vocabulary schema",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "title": "Vocabulary name",
                        "description": "Name of the vocabulary",
                        "type": "string",
                        "format": "string",
                        "nullable": false
                    },
                    "slug": {
                        "title": "Vocabulary slug",
                        "description": "Slug of the vocabulary",
                        "type": "string",
                        "format": "string",
                        "nullable": false
                    },
                    "description": {
                        "title": "Vocabulary description",
                        "description": "Description of the vocabulary",
                        "type": "string",
                        "format": "string",
                        "nullable": true
                    },
                    "createdAt": {
                        "title": "Created at",
                        "description": "Timestamp of creation",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-08-23 14:39:37",
                        "nullable": true
                    },
                    "updatedAt": {
                        "title": "Updated at",
                        "description": "Timestamp of last update",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-08-23 14:39:37",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "204": {
                "description": "No Content"
            },
            "401": {
                "description": "Unauthorized"
            },
            "403": {
                "description": "Forbidden"
            },
            "404": {
                "description": "Not Found"
            }
        },
        "parameters": {
            "search": {
                "name": "search",
                "in": "query",
                "description": "Search term",
                "required": false,
                "schema": {
                    "type": "string"
                }
            },
            "page": {
                "name": "page",
                "in": "query",
                "description": "Page number",
                "required": false,
                "schema": {
                    "type": "int",
                    "example": "1"
                }
            },
            "per_page": {
                "name": "per_page",
                "in": "query",
                "description": "Number of records per page (Default: 10)",
                "required": false,
                "schema": {
                    "type": "int",
                    "example": "10"
                }
            }
        }
    },
    "tags": [
        {
            "name": "terms",
            "description": "Vocabulary Terms"
        },
        {
            "name": "vocabularies",
            "description": "Vocabularies"
        }
    ]
}